requirements.txt
numpy
opencv-python
scipy
tyro
viser
open3d
pyyaml